Switch to scoped session
