+++ src/NServiceBus.Persistence.NonDurable.AcceptanceTests/NServiceBus.Persistence.NonDurable.AcceptanceTests.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="NServiceBus.AcceptanceTests.Sources" Version="8.0.0-beta.1" />
+    <PackageReference Include="NServiceBus.AcceptanceTests.Sources" Version="8.0.0-storagession.1" />
   </ItemGroup>
 
   <ItemGroup>
