+++ src/NServiceBus.Persistence.NonDurable.PersistenceTests/NServiceBus.Persistence.NonDurable.PersistenceTests.csproj
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="NServiceBus.PersistenceTests.Sources" Version="8.0.0-beta.1" />
+    <PackageReference Include="NServiceBus.PersistenceTests.Sources" Version="8.0.0-storagession.1" />
   </ItemGroup>
 
   <ItemGroup>
