+++ src/NServiceBus.Persistence.NonDurable/NServiceBus.Persistence.NonDurable.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup Label="Public dependencies">
-    <PackageReference Include="NServiceBus" Version="[8.0.0-beta.1, 9.0.0)" />
+    <PackageReference Include="NServiceBus" Version="8.0.0-storagession.1" />
   </ItemGroup>
 
   <ItemGroup Label="Private dependencies">
