--- src/NServiceBus.Persistence.NonDurable/NonDurableTransactionalSynchronizedStorageAdapter.cs
@@ -1,77 +1,0 @@
-namespace NServiceBus
-{
-    using System;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using System.Transactions;
-    using Extensibility;
-    using Outbox;
-    using Persistence;
-    using Transport;
-
-    class NonDurableTransactionalSynchronizedStorageAdapter : ISynchronizedStorageAdapter
-    {
-        public Task<ICompletableSynchronizedStorageSession> TryAdapt(IOutboxTransaction transaction, ContextBag context, CancellationToken cancellationToken = default)
-        {
-            if (transaction is NonDurableOutboxTransaction inMemOutboxTransaction)
-            {
-                ICompletableSynchronizedStorageSession session = new NonDurableSynchronizedStorageSession(inMemOutboxTransaction.Transaction);
-                return Task.FromResult(session);
-            }
-            return EmptyTask;
-        }
-
-        public Task<ICompletableSynchronizedStorageSession> TryAdapt(TransportTransaction transportTransaction, ContextBag context, CancellationToken cancellationToken = default)
-        {
-            if (transportTransaction.TryGet(out Transaction ambientTransaction))
-            {
-                var transaction = new NonDurableTransaction();
-                ICompletableSynchronizedStorageSession session = new NonDurableSynchronizedStorageSession(transaction);
-                ambientTransaction.EnlistVolatile(new EnlistmentNotification2(transaction), EnlistmentOptions.None);
-                return Task.FromResult(session);
-            }
-            return EmptyTask;
-        }
-
-        static readonly Task<ICompletableSynchronizedStorageSession> EmptyTask = Task.FromResult<ICompletableSynchronizedStorageSession>(null);
-
-        class EnlistmentNotification2 : IEnlistmentNotification
-        {
-            public EnlistmentNotification2(NonDurableTransaction transaction)
-            {
-                this.transaction = transaction;
-            }
-
-            public void Prepare(PreparingEnlistment preparingEnlistment)
-            {
-                try
-                {
-                    transaction.Commit();
-                    preparingEnlistment.Prepared();
-                }
-                catch (Exception ex)
-                {
-                    preparingEnlistment.ForceRollback(ex);
-                }
-            }
-
-            public void Commit(Enlistment enlistment)
-            {
-                enlistment.Done();
-            }
-
-            public void Rollback(Enlistment enlistment)
-            {
-                transaction.Rollback();
-                enlistment.Done();
-            }
-
-            public void InDoubt(Enlistment enlistment)
-            {
-                enlistment.Done();
-            }
-
-            readonly NonDurableTransaction transaction;
-        }
-    }
-}
